consolicate duplicate AuthMessage types
